+++ src/main/java/com/adit/backend/domain/user/repository/UserRepository.java
@@ -3,7 +3,6 @@ package com.adit.backend.domain.user.repository;
 import java.util.Optional;
 
 import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.data.jpa.repository.Query;
 import org.springframework.stereotype.Repository;
 
 import com.adit.backend.domain.user.entity.User;
@@ -13,7 +12,4 @@ public interface UserRepository extends JpaRepository<User, Long> {
 	Optional<User> findByEmail(String email);
 
 	boolean existsByNickname(String nickname);
-
-	@Query("SELECT u FROM User u LEFT JOIN FETCH u.token t WHERE t.accessToken = ?1")
-	Optional<User> findUserByToken_AccessToken(String accessToken);
 }
