+++ src/main/java/com/adit/backend/global/security/jwt/entity/Token.java
@@ -1,0 +1,17 @@
+package com.adit.backend.global.security.jwt.entity;
+
+import lombok.Builder;
+import lombok.Getter;
+
+@Getter
+public class Token {
+
+	private String accessToken;
+	private String refreshToken;
+
+	@Builder
+	public Token(String accessToken, String refreshToken) {
+		this.accessToken = accessToken;
+		this.refreshToken = refreshToken;
+	}
+}
